refactor job views
